Config Fix + Model correction
Updated table models
Fixed config data load

+++ models/recruiting.go
@@ -9,30 +9,30 @@ import (
 
 type RecruitingTeams []RecruitingTeam
 type RecruitingTeam struct {
-	Year   int    `json:"year"`
-	Rank   int    `json:"rank"`
-	Team   string `json:"team"`
-	Points int    `json:"points"`
+	Year   int     `json:"year"`
+	Rank   int     `json:"rank"`
+	Team   string  `json:"team"`
+	Points float64 `json:"points"`
 }
 
 type Recruits []Recruit
 type Recruit struct {
-	RecruitID     string `json:"id"`
-	AthleteID     string `json:"athleteId"`
-	RecruitType   string `json:"recruitType"`
-	Year          int    `json:"year"`
-	Ranking       int    `json:"ranking"`
-	Name          string `json:"name"`
-	School        string `json:"school"`
-	CommittedTo   string `json:"committedTo"`
-	Position      string `json:"position"`
-	Height        int    `json:"height"`
-	Weight        int    `json:"weight"`
-	Stars         int    `json:"stars"`
-	Rating        int    `json:"rating"`
-	City          string `json:"city"`
-	StateProvince string `json:"stateProvince"`
-	Country       string `json:"country"`
+	RecruitID     string  `json:"id"`
+	AthleteID     string  `json:"athleteId"`
+	RecruitType   string  `json:"recruitType"`
+	Year          int     `json:"year"`
+	Ranking       int     `json:"ranking"`
+	Name          string  `json:"name"`
+	School        string  `json:"school"`
+	CommittedTo   string  `json:"committedTo"`
+	Position      string  `json:"position"`
+	Height        int     `json:"height"`
+	Weight        int     `json:"weight"`
+	Stars         int     `json:"stars"`
+	Rating        float64 `json:"rating"`
+	City          string  `json:"city"`
+	StateProvince string  `json:"stateProvince"`
+	Country       string  `json:"country"`
 }
 
 func FetchAndInsertRecruits(year int) error {
