+++ models/seasonal/talent.go
@@ -9,9 +9,9 @@ import (
 
 type TeamTalent []Talent
 type Talent struct {
-	Year   int    `json:"year" gorm:"primaryKey"`
-	Team   string `json:"team" gorm:"primaryKey:size:100"`
-	Talent int    `json:"talent"`
+	Year   int     `json:"year" gorm:"primaryKey"`
+	Team   string  `json:"team" gorm:"primaryKey:size:100"`
+	Talent float64 `json:"talent"`
 }
 
 func FetchAndInsertTalent() error {
